+++ gokkan/src/pages/LayoutPage.tsx
@@ -2,11 +2,6 @@ import { Outlet } from 'react-router-dom';
 import styled from 'styled-components';
 import Header from '../components/Header/Header';
 
-const Main = styled.main`
-  margin-top: 32px;
-  padding: 0 16px;
-`;
-
 export default function LayoutPage() {
   return (
     <div>
@@ -17,3 +12,8 @@ export default function LayoutPage() {
     </div>
   );
 }
+
+const Main = styled.main`
+  margin-top: 32px;
+  padding: 0 16px;
+`;
